requirements.txt
# Frontend
streamlit>=1.28.0

# Backend
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
pdfplumber>=0.10.0

# NLP and Embeddings
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.2.0

# Utilities
numpy>=1.26.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Text Processing
nltk>=3.8.1
